api/requirements.txt
click==8.0.4
colorama==0.4.4
Faker==4.0.3
Flask==2.0.3
Flask-PyMongo==2.3.0
itsdangerous==2.1.1
Jinja2==3.0.3
kafka-python==2.0.2
MarkupSafe==2.1.1
PyJWT==2.3.0
pymongo==4.0.2
python-dateutil==2.8.2
six==1.16.0
text-unidecode==1.3
Werkzeug==2.0.3